Develop

+++ app/src/main/java/com/albuquerque/cryptoe_wallet/app/model/entity/UserEntity.kt
@@ -2,11 +2,12 @@ package com.albuquerque.cryptoe_wallet.app.model.entity
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import java.math.BigDecimal
 
 @Entity(tableName = "user")
 data class UserEntity(
     @PrimaryKey val email: String,
     val password: String,
     val fullName: String,
-    val balance: Double = 100000.0
+    val balance: BigDecimal = BigDecimal(100000.0)
 )
